Identify missing countries
Not all of the "required" countries have embassy listings on the
website. These have been identified so that they can be dealt with
manually.

Other countries were "lost in translation", because their names
didn't match up with those in the database or because multiple
countries used the same Demonym. These have been added to the
"extra_demonyms" list so that they're dealt with.